SS-16 | Implementação escuta de canais
Implementação de escuta de canais completamente funcional, onde o bot consegue ler qualquer mensagem que é posta no canal do slack.
